Metrics.width, Metrics.height 를 이용하여 view 의 폭과 높이를 구하도록 한다.

+++ projects/SampleGame01/app/src/main/java/kr/ac/tukorea/ge/sgp02/s12345678/samplegame01/Fighter.java
@@ -21,7 +21,7 @@ public class Fighter implements GameObject {
     private static Bitmap bitmap;
     private static Rect srcRect = new Rect();
 
-    public Fighter(int x, int y) {
+    public Fighter(float x, float y) {
         this.x = x;
         this.y = y;
         this.tx = x;
